Assistant checkpoint: Update download-ics endpoint to receive events from client
Assistant generated file changes:
- routes.py: Update download-ics endpoint to accept events from request
- static/js/app.js: Update download-ics request to send events data

---

User prompt:

on a browser on a mac in safari, I am getting this when I try to download

{
  "error_type": "no_events",
  "success": false,
  "user_message": "No events found to download"
}

But this doesn't seem to happen if I try from the safari iOS browser.

What could be causing this? Can you double check that on download we are passing all the events properly into the backend since they're only stored in the client?